量房费改类型

+++ service/src/main/java/cn/thinkfree/service/platform/designer/CreatePayOrderService.java
@@ -9,5 +9,5 @@ public interface CreatePayOrderService {
      * 创建量房订单
      * @param projectNo
      */
-    void createVolumeRoomPay(String projectNo, Integer appointmentAmount);
+    void createVolumeRoomPay(String projectNo, Float appointmentAmount);
 }
